requirements.txt
# Semantic Similarity Rating - Python Package Dependencies
# Python >= 3.9, <= 3.12 recommended (3.13 not fully supported by torch)

# Core packages
polars>=0.18.0
numpy>=1.24.0
pandas>=2.0.0

# LLM API
openai>=1.0.0

# Streamlit dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Scientific computing
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
